requirements.txt
garminconnect
google-api-python-client
google-auth-httplib2
google-auth-oauthlib